requirements.txt
alipay-sdk-python==3.6.733
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
daphne==4.0.0
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
hiredis==2.2.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
Pillow==10.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.8.0
pyOpenSSL==23.2.0
pytz==2023.3
redis==5.0.0
requests==2.31.0
rsa==4.9
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
supervisor==4.2.5
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
zope.interface==6.0
